Remove broken dialpad reveal animation
After migrating to Androidx, the dialpad reveal animation breaks when adding this fade animation to the search fragment. Removing this animation doesn't change the appearance of the transition and fixes the issue where the dialpad will appear, disappear, and then scroll in when you invoke it for the first time.